requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.10.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0 
asn1crypto==1.5.1
attrs==25.4.0
bcrypt==5.0.0
boto3==1.42.4
botocore==1.42.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.124.0
frozenlist==1.8.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
greenlet==3.2.4
h11==0.14.0
# httpcore==0.16.3
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
oauthlib==3.3.1
orjson==3.11.5
passlib==1.7.4
pg8000==1.31.5
propcache==0.4.1
psycopg2-binary==2.9.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.6.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy>=2.0.32
starlette==0.49.3
twilio==7.16.0
typing_extensions==4.15.0
ujson==5.11.0
urllib3==1.26.18
uvicorn==0.22.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0

ecdsa==0.19.1

passlib==1.7.4

psycopg2-binary==2.9.6

python-jose==3.5.0
razorpay==1.4.2

scramp==1.4.6
typing-inspection==0.4.2
tzdata==2025.3
apscheduler==3.11.2